First try at creating a new type of filesystem
What has changed? I made a more ordinary fs that depends more on sshfs
than on SFTP (that version is obsolete), because with sshfs you can use
ordinary os and path things in Python and still you can have your files
stored in an ordinary fashion.

But the most important thing is that the new FS works with a central
file index.